+++ kernel/drivers/chrdev.c
@@ -25,12 +25,12 @@ int register_char_device(char_device_t *device) {
     return 0;
 }
 
-int char_read(char_device_t device, uint64_t offset, uint8_t *buffer, uint64_t size) {
+int64_t char_read(char_device_t device, uint64_t offset, uint8_t *buffer, uint64_t size) {
     if (device.major_number == 0 || !char_drivers[device.major_number].present) return -ENODEV;
     return char_drivers[device.major_number].read(device.minor_number, offset, buffer, size);
 }
 
-int char_write(char_device_t device, uint64_t offset, const uint8_t *buffer, uint64_t size) {
+int64_t char_write(char_device_t device, uint64_t offset, const uint8_t *buffer, uint64_t size) {
     if (device.major_number == 0 || !char_drivers[device.major_number].present) return -ENODEV;
     return char_drivers[device.major_number].write(device.minor_number, offset, buffer, size);
 }
